core: add nm_config_data_get_device_config_by_pllink() to lookup per-device config
... by platform link.

One caveat is that without having a NMDevice instance, matching by
several paramters won't work. Like, matching against the driver would
require us to look it up via ethtool. When having an NMDevice instance,
the driver is cached there, it's unclear we want to call ethtool for
lookup in this case -- though it could be done.

For other options, it's more complicated. Like, the type basically
depends on the NMDevice class. Usually that also works without a
netdev known to kernel (like bluetooth).

The inconsistency that certain matches are not implemented is ugly
indeed. But the effect is as if the spec doesn't match.