build: fix libnm_linking test
The `libnm_linking` test that belongs to the libnm-util's general
tests is failing because the test is not able to find the
`test-libnm-linking` binary, which is executed as a child process.

The problem lies to the `BUILD_DIR` macro definition which is
used to set the place to find the binary, and is wrongly defined
with the source directory.

This patch changes its value to the build directory that fixes
the problem.

https://mail.gnome.org/archives/networkmanager-list/2017-December/msg00049.html