src/tests: fix test_nm_utils_kill_child() under meson
meson spawns the tests in a way that the test process is
a session leader. Since the test wants to create a new
process group to kill a group of processes that it starts,
it failed.

  $ meson test -C build test-general-with-expect

The test would have succeed when wrapping the test for example
by strace:

  $ meson test -C build --wrap='strace' test-general-with-expect

Fix that, by forking once more.