iwd: fix coding style to use curly braces for multi-line if()
Our convention is that when the body of an if() or for() spawns
more then one line, then it needs curly braces. If it's only one
line, it should have no curly braces. The latter part seems sometimes
a bit inconvenient, because changing

  if (some_condition)
      do_something ();

gets change to

  if (some_condition) {
      do_something ();
      do_something_else ();
  }

the diff shows 3 lines changed, although really only one changed.

But well, that's how it is...