remove --run-from-build-dir option from NetworkManager
It used to be possible to successfully run NetworkManager from the build
directory at least using `--run-from-build-dir`. The code is relatively
complicated for that simple task and doesn't even achieve the results.

For example the path to device plugins is hardcoded and the actual device
plugins are in their own different build directories. There is no trivial
way to fix that.

In future we should probably implement running from build directory using a
set of environment variables with search paths and other tweaks together
with `NetworkManager` and `nmcli` shell wrappers that would make it all
work. In addition to plugin and helper search paths there appears to be a
problem with the dbus and polkit configurations in the running system. This
seems to be a drawback related to the usage of D-Bus for all communication.