dhcp: remove dhcpcd plugin
The dhcpcd plugin doesn't support IPv6 and there is no good reason to use
it with NetworkManager. The power of dhcpcd is when it is used as a
standalone IPv4/IPv6 configuration daemon rather than a DHCP helper for one
of the big network configuration services. The default DHCP plugin for
NetworkManager is now dhclient even on Gentoo and we are eventually going
to switch away from that as well.

Removing dhcpcd plugin ensures that we don't have additional maintainance
burden with a plugin that isn't even fully working and nobody was willing
to fix it for many years.