libnm: rework json wrapper to expose VTable and synchronize load
Rework the code, so that loading is threadsafe.

Also, expose a NMJsonVt virtual table, that can be used instead
of the plain json function. It could be used to minimize the overhead
of the wrapper functions.

Downsides are, that the function types must now be explicitly
specified in code, which is more work to get done. However, it's
not dangerious, as the compiler enforces that the signature is
correct.

Later, this will expand to not to make use of any symbols from
the jansson.h header.